Modify structure to keep right dependencies.
because this app in :dev/:test is either a (compile-time) tool and the demo (3gpp),
the demo is split into erlang part (.asn compiled to .hrl/.erl/.asn1db)
and elixir part (all rest stuff that use compile-time goodies of this app).

in real projects, where this app will be a clear dependency,
any other app using this app would be undivided.