Add optional reck2v argument to ASN.RTT.from_map/2
* reck2v(record_type, field_name) -> value | nil

If the value of mandatory field is :undefined, reck2v can provide the value.
reck2v may raise exception or return nil. In the later case from_map/2 will raise.

=== http://erlang.org/doc/apps/asn1/asn1_getting_started.html

Extensibility for SEQUENCE and SET

When a SEQUENCE or SET contains an extension marker and extension components as the following, the type can get more components in newer versions of the ASN.1 spec:

SExt ::= SEQUENCE {
           a INTEGER,
           ...,
           b BOOLEAN }

In this case it has got a new component b. Thus, incoming messages that are decoded can have more or fever components than this one.

The component b is treated as an original component when encoding a message. In this case, as it is not an optional element, it must be encoded.

During decoding, the b field of the record gets the decoded value of the b component, if present, otherwise the value asn1_NOVALUE.